checkbook request implemented
